separation of request validation and matching
