+++ hmt/serve/mock/request_processor.py
@@ -46,9 +46,9 @@ class RequestProcessor:
             timestamp=None,
         )
 
-    def _match_response(self, spec: OpenAPISpecification, request: Request):
+    def _match_response(self, pathname: str, spec: OpenAPISpecification, request: Request):
         try:
-            return self._faker.process(spec, request)
+            return self._faker.process(pathname, spec, request)
         except FakerException as e:
             return self.match_error(str(e), request)
 
@@ -73,27 +73,19 @@ class RequestProcessor:
             return maybe_security_response
 
         logger.debug("Matching to openapi")
-        match = match_request_to_openapi(request, specs)
+        pathname, spec = match_request_to_openapi(request, specs)
         logger.debug("Finished matching to openapi")
 
-        if len(match) == 0:
+        if pathname is None:
             return self._callback_manager(
                 request,
                 self.match_error(
-                    "Could not find an open API schema for the host %s." % request.host,
+                    "Could not find an open API schema for the host {} and the path {}".format(request.host, request.path),
                     request,
                 ),
                 self._mock_data_store.default,
             )
 
-        spec = random.choice(match)
-        if spec.api.paths is None or len(spec.api.paths.items()) == 0:
-            path_error = "Could not find a path %s on hostname %s." % (
-                request.path,
-                request.host,
-            )
-            return self.match_error(path_error, request)
-
         storage = self._mock_data_store[spec.source]
-        response = self._match_response(spec, request)
+        response = self._match_response(pathname, spec, request)
         return self._callback_manager(request, response, storage)
