+++ hmt/serve/mock/specs.py
@@ -5,6 +5,7 @@ from typing import Sequence, Union, Any
 
 import requests
 import yaml
+from hmt.serve.mock.refs import make_definitions_from_spec
 from openapi_typed_2 import OpenAPIObject, convert_to_openapi
 from requests.exceptions import RequestException
 
@@ -15,7 +16,7 @@ class OpenAPISpecification:
 
     api: OpenAPIObject
     source: str
-
+    definitions: Any
 
 def load_spec(spec_source: str, is_http: bool) -> OpenAPISpecification:
     spec_text: str
@@ -29,11 +30,12 @@ def load_spec(spec_source: str, is_http: bool) -> OpenAPISpecification:
         with open(spec_source, encoding="utf8") as spec_file:
             spec_text = spec_file.read()
 
-    return OpenAPISpecification(
-        convert_to_openapi(
+    spec =  convert_to_openapi(
             (json.loads if spec_source.endswith("json") else yaml.safe_load)(spec_text)
-        ),
-        spec_source,
+        )
+    return OpenAPISpecification(
+        spec,
+        spec_source, make_definitions_from_spec(spec)
     )
