+++ tests/serve/mock/test_matcher_helper_functions.py
@@ -11,7 +11,7 @@ from openapi_typed_2 import (
     convert_to_Schema,
 )
 
-from hmt.serve.mock.matcher import (
+from hmt.serve.mock.request_validation import (
     get_path_item_with_method,
     match_urls,
     matches,
